# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Adare General Hospital Platform API",
    "version": "1.0.0",
    "description": "REST API for the Adare General Hospital digital platform (Hawassa, Sidama Region, Ethiopia). All responses use the envelope { ok, data, message } or { ok:false, code, error }. Protected endpoints require Authorization: Bearer <access_token>; refresh tokens rotate via HTTP-only cookie on /api/auth."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
    "schemas": {
...
      "AppointmentCreate": {
        "type": "object",
        "required": ["patient_name", "phone", "preferred_date"],
        "properties": {
          "patient_name": { "type": "string" }, "phone": { "type": "string", "example": "0911223344" },
          "email": { "type": "string" }, "gender": { "enum": ["male", "female", "other"] },
          "date_of_birth": { "type": "string", "format": "date" },
          "department_id": { "type": "integer" }, "doctor_id": { "type": "integer" }, "service_id": { "type": "integer" },
          "preferred_date": { "type": "string", "format": "date" }, "preferred_time": { "type": "string" },
          "reason": { "type": "string" }, "is_emergency": { "type": "boolean" },
          "insurance_type": { "enum": ["none", "cbhi", "private", "other"] }, "notes": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/login": { "post": { "summary": "Staff or patient login", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string" } } } } } }, "responses": { "200": { "description": "access_token + user" }, "401": { "description": "Invalid credentials" }, "429": { "description": "Locked / rate-limited" } } } },
    "/auth/register": { "post": { "summary": "Patient self-registration", "responses": { "201": { "description": "Account + patient number" }, "409": { "description": "Duplicate phone" } } } },
    "/auth/refresh": { "post": { "summary": "Rotate refresh token (cookie) → new access token", "responses": { "200": { "description": "New access token" }, "401": { "description": "Expired" } } } },
    "/auth/logout": { "post": { "security": [{ "bearerAuth": [] }], "summary": "Sign out and revoke refresh token", "responses": { "200": { "description": "OK" } } } },
    "/auth/me": { "get": { "security": [{ "bearerAuth": [] }], "summary": "Current session", "responses": { "200": { "description": "User (+patient when role=patient)" } } } },
    "/auth/change-password": { "post": { "security": [{ "bearerAuth": [] }], "summary": "Change own password", "responses": { "200": { "description": "OK" } } } },
...
    "/appointments": {
      "post": { "summary": "Book appointment → AGH-APT reference", "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppointmentCreate" } } } }, "responses": { "201": { "description": "PENDING appointment" }, "409": { "description": "Duplicate" }, "422": { "description": "Validation" } } },
      "get": { "security": [{ "bearerAuth": [] }], "summary": "Staff appointment list (filters: status, date, department_id, doctor_id, q)", "responses": { "200": { "description": "Appointments" }, "403": { "description": "Role not allowed" } } }
    },
    "/appointments/status": { "get": { "summary": "Public status lookup (reference + phone)", "responses": { "200": { "description": "Status" }, "404": { "description": "Not found" } } } },
    "/appointments/{reference}": { "patch": { "security": [{ "bearerAuth": [] }], "summary": "Workflow action: confirm|reject|reschedule|checkin|start|complete|cancel|noshow", "responses": { "200": { "description": "New status" }, "409": { "description": "Invalid transition" } } } },
    "/appointments/{reference}/history": { "get": { "security": [{ "bearerAuth": [] }], "summary": "Status history", "responses": { "200": { "description": "Events" } } } },
...
  }
}